requirements.txt
torch==0.4.1
tqdm==4.19.5
pandas==0.22.0
numpy==1.13.1
torchvision==0.2.0
scipy==1.0.0
Pillow==5.3.0
progressbar2==3.38.0
